spark/app/requirements.txt
kafka-python-ng
requests
pandas